Make better use of ES6 modules and rollup
This helped drop a few kilobytes from the file size
