+++ src/linkify/core/scanner.js
@@ -7,56 +7,72 @@
 	@main scanner
 */
 
-import {text as TOKENS} from './tokens';
 import {CharacterState as State, stateify} from './state';
+import * as TOKENS from './tokens/text';
+import {
+	DOMAIN,
+	LOCALHOST,
+	NUM,
+	PROTOCOL,
+	TLD,
+	WS,
+	AT,
+	DOT,
+	PLUS,
+	POUND,
+	QUERY,
+	SLASH,
+	UNDERSCORE,
+	COLON,
+	OPENBRACE,
+	OPENBRACKET,
+	OPENPAREN,
+	CLOSEBRACE,
+	CLOSEBRACKET,
+	CLOSEPAREN,
+	PUNCTUATION,
+	NL,
+	SYM
+} from './tokens/text';
 
 const tlds = __TLDS__; // macro, see gulpfile.js
 
-const NUM = '0123456789'.split('');
+const NUMBERS = '0123456789'.split('');
 const ALPHANUM = '0123456789abcdefghijklmnopqrstuvwxyz'.split('');
 const WHITESPACE = [' ', '\f', '\r', '\t', '\v']; // excluding line breaks
-const COLON = ':';
 
 let domainStates = []; // states that jump to DOMAIN on /[a-z0-9]/
 let makeState = (tokenClass) => new State(tokenClass);
 
-// Frequently used tokens
-const T_DOMAIN		= TOKENS.DOMAIN;
-const T_LOCALHOST	= TOKENS.LOCALHOST;
-const T_NUM			= TOKENS.NUM;
-const T_PROTOCOL	= TOKENS.PROTOCOL;
-const T_TLD			= TOKENS.TLD;
-const T_WS			= TOKENS.WS;
-
 // Frequently used states
 const S_START			= makeState();
-const S_NUM				= makeState(T_NUM);
-const S_DOMAIN			= makeState(T_DOMAIN);
+const S_NUM				= makeState(NUM);
+const S_DOMAIN			= makeState(DOMAIN);
 const S_DOMAIN_HYPHEN	= makeState(); // domain followed by 1 or more hyphen characters
-const S_WS				= makeState(T_WS);
+const S_WS				= makeState(WS);
 
 // States for special URL symbols
 S_START
-.on('@', makeState(TOKENS.AT))
-.on('.', makeState(TOKENS.DOT))
-.on('+', makeState(TOKENS.PLUS))
-.on('#', makeState(TOKENS.POUND))
-.on('?', makeState(TOKENS.QUERY))
-.on('/', makeState(TOKENS.SLASH))
-.on('_', makeState(TOKENS.UNDERSCORE))
-.on(COLON, makeState(TOKENS.COLON))
-.on('{', makeState(TOKENS.OPENBRACE))
-.on('[', makeState(TOKENS.OPENBRACKET))
-.on('(', makeState(TOKENS.OPENPAREN))
-.on('}', makeState(TOKENS.CLOSEBRACE))
-.on(']', makeState(TOKENS.CLOSEBRACKET))
-.on(')', makeState(TOKENS.CLOSEPAREN))
-.on([',', ';', '!', '"'], makeState(TOKENS.PUNCTUATION));
+.on('@', makeState(AT))
+.on('.', makeState(DOT))
+.on('+', makeState(PLUS))
+.on('#', makeState(POUND))
+.on('?', makeState(QUERY))
+.on('/', makeState(SLASH))
+.on('_', makeState(UNDERSCORE))
+.on(':', makeState(COLON))
+.on('{', makeState(OPENBRACE))
+.on('[', makeState(OPENBRACKET))
+.on('(', makeState(OPENPAREN))
+.on('}', makeState(CLOSEBRACE))
+.on(']', makeState(CLOSEBRACKET))
+.on(')', makeState(CLOSEPAREN))
+.on([',', ';', '!', '"'], makeState(PUNCTUATION));
 
 // Whitespace jumps
 // Tokens of only non-newline whitespace are arbitrarily long
 S_START
-.on('\n', makeState(TOKENS.NL))
+.on('\n', makeState(NL))
 .on(WHITESPACE, S_WS);
 
 // If any whitespace except newline, more whitespace!
@@ -65,14 +81,14 @@ S_WS.on(WHITESPACE, S_WS);
 // Generates states for top-level domains
 // Note that this is most accurate when tlds are in alphabetical order
 for (let i = 0; i < tlds.length; i++) {
-	let newStates = stateify(tlds[i], S_START, T_TLD, T_DOMAIN);
+	let newStates = stateify(tlds[i], S_START, TLD, DOMAIN);
 	domainStates.push.apply(domainStates, newStates);
 }
 
 // Collect the states generated by different protocls
-let partialProtocolFileStates	= stateify('file', S_START, T_DOMAIN, T_DOMAIN);
-let partialProtocolFtpStates	= stateify('ftp', S_START, T_DOMAIN, T_DOMAIN);
-let partialProtocolHttpStates	= stateify('http', S_START, T_DOMAIN, T_DOMAIN);
+let partialProtocolFileStates = stateify('file', S_START, DOMAIN, DOMAIN);
+let partialProtocolFtpStates = stateify('ftp', S_START, DOMAIN, DOMAIN);
+let partialProtocolHttpStates = stateify('http', S_START, DOMAIN, DOMAIN);
 
 // Add the states to the array of DOMAINeric states
 domainStates.push.apply(domainStates, partialProtocolFileStates);
@@ -80,38 +96,38 @@ domainStates.push.apply(domainStates, partialProtocolFtpStates);
 domainStates.push.apply(domainStates, partialProtocolHttpStates);
 
 // Protocol states
-let S_PROTOCOL_FILE		= partialProtocolFileStates.pop();
-let S_PROTOCOL_FTP		= partialProtocolFtpStates.pop();
-let S_PROTOCOL_HTTP		= partialProtocolHttpStates.pop();
-let S_PROTOCOL_SECURE	= makeState(T_DOMAIN);
-let S_FULL_PROTOCOL		= makeState(T_PROTOCOL); // Full protocol ends with COLON
+let S_PROTOCOL_FILE = partialProtocolFileStates.pop();
+let S_PROTOCOL_FTP = partialProtocolFtpStates.pop();
+let S_PROTOCOL_HTTP = partialProtocolHttpStates.pop();
+let S_PROTOCOL_SECURE = makeState(DOMAIN);
+let S_FULL_PROTOCOL = makeState(PROTOCOL); // Full protocol ends with COLON
 
 // Secure protocols (end with 's')
 S_PROTOCOL_FTP
 .on('s', S_PROTOCOL_SECURE)
-.on(COLON, S_FULL_PROTOCOL);
+.on(':', S_FULL_PROTOCOL);
 
 S_PROTOCOL_HTTP
 .on('s', S_PROTOCOL_SECURE)
-.on(COLON, S_FULL_PROTOCOL);
+.on(':', S_FULL_PROTOCOL);
 
 domainStates.push(S_PROTOCOL_SECURE);
 
 // Become protocol tokens after a COLON
-S_PROTOCOL_FILE.on(COLON, S_FULL_PROTOCOL);
-S_PROTOCOL_SECURE.on(COLON, S_FULL_PROTOCOL);
+S_PROTOCOL_FILE.on(':', S_FULL_PROTOCOL);
+S_PROTOCOL_SECURE.on(':', S_FULL_PROTOCOL);
 
 // Localhost
-let partialLocalhostStates = stateify('localhost', S_START, T_LOCALHOST, T_DOMAIN);
+let partialLocalhostStates = stateify('localhost', S_START, LOCALHOST, DOMAIN);
 domainStates.push.apply(domainStates, partialLocalhostStates);
 
 // Everything else
 // DOMAINs make more DOMAINs
 // Number and character transitions
-S_START.on(NUM, S_NUM);
+S_START.on(NUMBERS, S_NUM);
 S_NUM
 .on('-', S_DOMAIN_HYPHEN)
-.on(NUM, S_NUM)
+.on(NUMBERS, S_NUM)
 .on(ALPHANUM, S_DOMAIN); // number becomes DOMAIN
 
 S_DOMAIN
@@ -127,11 +143,11 @@ for (let i = 0; i < domainStates.length; i++) {
 
 S_DOMAIN_HYPHEN
 .on('-', S_DOMAIN_HYPHEN)
-.on(NUM, S_DOMAIN)
+.on(NUMBERS, S_DOMAIN)
 .on(ALPHANUM, S_DOMAIN);
 
 // Set default transition
-S_START.defaultTransition = makeState(TOKENS.SYM);
+S_START.defaultTransition = makeState(SYM);
 
 /**
 	Given a string, returns an array of TOKEN instances representing the
