+++ src/linkify/core/tokens/create-token-class.js
@@ -1,0 +1,9 @@
+function createTokenClass() {
+	return function (value) {
+		if (value) {
+			this.v = value;
+		}
+	};
+}
+
+export {createTokenClass};
